requirements.txt
flask==2.3.3
requests==2.31.0
kafka-python==2.0.2
